+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## 0.2.1
+
+* Add `thousandSeparator` property. E.g. `thousandSeparator: ','` turns `1000` into `1,000`.
+* Add `decimalSeparator` property to custom the decimal point for certain locale.
+
 ## 0.2.0
 
 * Add `wholeDigits` property. E.g. `wholeDigits: 4` turns `48` into `0048`.
-
 
 ## 0.1.2